Make link types configurable via editor.config.linkTypes
The list of link types in the Link Info tab of the Link dialog
can now be configured via the configuration string 'linkTypes',
which contains a comma-separated list of link types. The default
value is 'url,anchor,email', so by default the Link dialog will
show the same options as before. The option selected by default
is the first item in the comma-separated list.