CMS-10712 Enable customization of extraAllowedContent rules
The string version of 'extraAllowedContent' cannot be extended
because appending an extra string of rules to it will add a comma
in between (the logic in Json#append for concatenating string
properties with a comma). Solution: use the object version of
'extraAllowedContent' instead so projects can overlay extra rules
into it, and/or append to the existing allowed attributes (which are
comma-separated).