CHANNELMGR-1309 Include icons of codemirror plugin in distribution
The codemirror plugin overrides the sourcedialog completely. It
adds four buttons to the top. Yet those button use individual icon
files (as a CSS background image) and not icons in the sprite
generated by CKBuilder. As a result the buttons don't have icons
when the 'optimized' Hippo CKEditor distribution is used.

This seems a bug in the codemirror plugin. As a quick fix we
include the icons in the optimized build so the buttons show icons
again.