CMS-10944 Disable hippopicker buttons when their markup is not allowed
The commands of the 'internal link' and 'image' buttons need the ACF
rules too, otherwise the buttons are not toggled properly within
sections of a widget.

The requiredContent rules for the 'internal link' and 'image' buttons
used exclamation marks, which removed both buttons when content from
either one was not allowed. For example, when the 'img' was not allowed,
the internal link buttons got removed too. Not using exclamation marks
fixed this.

(cherry picked from commit 7fddce25798b7b72a475d95955b1f3f4553751e6)