CMS-11021 Parse image URL before fetching it
The URL of the image prefetched by the image2 plugin should always be a
valid, absolute URL. However, while entering an absolute URL the value
can be interpreted as a relative URL (e.g. "http", the start of an
absolute URL, will be interpreted as a relative URL by the browser).
Such relative URLs should be ignored. The URL constructor throws an
error for illegal and relative URLs, so that's the simple yet effective
filter.

Providing this fix upstream is hard because officially CKEditor
still supports IE. However, IE does not support the URL constructor.